// client/stocklinx app/src/redux/assetProduct/constant.ts
export enum assetProductConst {
  FETCH_ASSETPRODUCTS_REQUEST = "FETCH_ASSETPRODUCTS_REQUEST",
  FETCH_ASSETPRODUCTS_SUCCESS = "FETCH_ASSETPRODUCTS_SUCCESS",
  FETCH_ASSETPRODUCTS_FAILURE = "FETCH_ASSETPRODUCTS_FAILURE",
  FETCH_ASSETPRODUCT_REQUEST = "FETCH_ASSETPRODUCT_REQUEST",
  FETCH_ASSETPRODUCT_SUCCESS = "FETCH_ASSETPRODUCT_SUCCESS",
  FETCH_ASSETPRODUCT_FAILURE = "FETCH_ASSETPRODUCT_FAILURE",
  FETCH_ASSETPRODUCT_COUNTS_REQUEST = "FETCH_ASSETPRODUCT_COUNTS_REQUEST",
  FETCH_ASSETPRODUCT_COUNTS_SUCCESS = "FETCH_ASSETPRODUCT_COUNTS_SUCCESS",
  FETCH_ASSETPRODUCT_COUNTS_FAILURE = "FETCH_ASSETPRODUCT_COUNTS_FAILURE",
  CREATE_ASSETPRODUCT_REQUEST = "CREATE_ASSETPRODUCT_REQUEST",
  CREATE_ASSETPRODUCT_SUCCESS = "CREATE_ASSETPRODUCT_SUCCESS",
  CREATE_ASSETPRODUCT_FAILURE = "CREATE_ASSETPRODUCT_FAILURE",
  CREATE_RANGE_ASSETPRODUCT_REQUEST = "CREATE_RANGE_ASSETPRODUCT_REQUEST",
  CREATE_RANGE_ASSETPRODUCT_SUCCESS = "CREATE_RANGE_ASSETPRODUCT_SUCCESS",
  CREATE_RANGE_ASSETPRODUCT_FAILURE = "CREATE_RANGE_ASSETPRODUCT_FAILURE",
  UPDATE_ASSETPRODUCT_REQUEST = "UPDATE_ASSETPRODUCT_REQUEST",
  UPDATE_ASSETPRODUCT_SUCCESS = "UPDATE_ASSETPRODUCT_SUCCESS",
  UPDATE_ASSETPRODUCT_FAILURE = "UPDATE_ASSETPRODUCT_FAILURE",
  REMOVE_ASSETPRODUCT_REQUEST = "REMOVE_ASSETPRODUCT_REQUEST",
  REMOVE_ASSETPRODUCT_SUCCESS = "REMOVE_ASSETPRODUCT_SUCCESS",
  REMOVE_ASSETPRODUCT_FAILURE = "REMOVE_ASSETPRODUCT_FAILURE",
  REMOVE_RANGE_ASSETPRODUCT_REQUEST = "REMOVE_RANGE_ASSETPRODUCT_REQUEST",
  REMOVE_RANGE_ASSETPRODUCT_SUCCESS = "REMOVE_RANGE_ASSETPRODUCT_SUCCESS",
  REMOVE_RANGE_ASSETPRODUCT_FAILURE = "REMOVE_RANGE_ASSETPRODUCT_FAILURE",
  SET_ASSETPRODUCT = "SET_ASSETPRODUCT",
  CLEAR_ASSETPRODUCT = "CLEAR_ASSETPRODUCT",
  SET_ASSETPRODUCTS = "SET_ASSETPRODUCTS",
  CLEAR_ASSETPRODUCTS = "CLEAR_ASSETPRODUCTS",
}